% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{as.numeric,Dice-method}
\alias{as.numeric,Dice-method}
\title{Roll a dice}
\usage{
\S4method{as.numeric}{Dice}(x, ...)
}
\arguments{
\item{x}{A Dice object}

\item{...}{Not used}
}
\description{
Roll a dice
}
