% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{as.numeric.Dice}
\alias{as.numeric.Dice}
\title{Coerce Dice object to numeric scalar by simulating a roll}
\usage{
\method{as.numeric}{Dice}(x, ...)
}
\arguments{
\item{x}{A Dice object.}

\item{...}{Not used.}
}
\description{
Coerce Dice object to numeric scalar by simulating a roll
}
