% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attack.R
\name{attack-check}
\alias{attack-check}
\alias{attack_hit}
\alias{attack_miss}
\alias{attack_critical_hit}
\alias{attack_critical_miss}
\alias{check_pass}
\alias{check_fail}
\alias{check_critical_pass}
\alias{check_critical_fail}
\title{Get the probability of passing a check or hitting an attack}
\usage{
attack_hit(roll, ac)

attack_miss(roll, ac)

attack_critical_hit(roll)

attack_critical_miss(roll)

check_pass(roll, dc)

check_fail(roll, dc)

check_critical_pass(roll)

check_critical_fail(roll)
}
\arguments{
\item{roll}{A roll expression (e.g., \code{2 * d6 + 5}) or a data frame returned
by \code{\link[=r]{r()}}.}

\item{ac}{The armor class to beat for an attack.}

\item{dc}{The dificulty class to beat for a skill check.}
}
\value{
A numeric scalar.
}
\description{
These are simple functions that calculate probabilities of hits/misses and
passes/fails. They hide the complexities of the \link{roll} family of functions
so users unfamiliar with R's d/p/q/r notation can get quickly up and running
with the package.
}
\details{
This family of functions takes a roll expression (i.e., an arithmetic
expression involving dice) and calculates the probability of either
hitting/missing an attack or passing/failing a check. There are also
functions to calculate the probability of a critical hit/miss and of a
critical pass/fail.

Note that a critical success is considered to occur when the given die
rolls its highest possible value and that a critical fumble occurs when
opposite happens.

For more details on roll expressions, see \code{\link[=r]{r()}} and the \link{Dice} S4 class.
}
\examples{
# Probability of d20 + 8 passing a DC 15 skill check
check_pass(d20 + 8, 15)

# Probability of d20 + 8 missing an AC 15
attack_miss(d20 + 8, 15)

# Probability of disadvantaged roll critting
attack_critical_hit(min(d20, d20))

}
\seealso{
\link{roll}, \code{\link[=r]{r()}}, \link{Dice}.
}
