% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{c,Dice-method}
\alias{c,Dice-method}
\title{Combine dice}
\usage{
\S4method{c}{Dice}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{A Dice object}

\item{...}{Objects to concatenate}

\item{recursive}{Recursively descend}
}
\description{
Combine dice
}
