% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\alias{check_pass}
\alias{check_fail}
\title{Get the probability of passing or failing a check}
\usage{
check_pass(roll, dc)

check_fail(roll, dc)
}
\arguments{
\item{roll}{A roll expression (e.g., \code{2 * d6 + 5}) or a data frame returned
by \code{\link[=r]{r()}}.}

\item{dc}{The dificulty class to beat for a skill check. Conversely, the
armor class to beat for an attack.}
}
\value{
A numeric scalar.
}
\description{
These are simple functions that calculate probabilities of hits/misses and
passes/fails. They hide the complexities of the \link{roll} family of functions
so users unfamiliar with R's d/p/q/r notation can get quickly up and running
with the package.
}
\details{
This family of functions takes a roll expression (i.e., an arithmetic
expression involving dice) and calculates the probability of either
hitting/missing an attack or passing/failing a check. Since attacks and
checks work in the same way (i.e., roll a value higher than or equal to a
certain threshold), there are no \verb{attack_*()} functions.

For more details on roll expressions, see \code{\link[=r]{r()}} and the \link{Dice} S4 class.
}
\examples{
# Probability of d20 + 8 passing a DC 15 skill check
check_pass(d20 + 8, 15)

# Probability of d20 + 8 missing an AC 15 attack
check_fail(d20 + 8, 15)

}
\seealso{
\link{roll}, \code{\link[=r]{r()}}, \link{Dice}.
}
