% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dc}
\alias{check_dc}
\title{Get the DC of a check given its success/failure probability}
\usage{
check_dc(roll, p, pass = TRUE)
}
\arguments{
\item{roll}{A roll expression (e.g., \code{2 * d6 + 5}) or a data frame returned
by \code{\link[=r]{r()}}.}

\item{p}{The probability of success/failure of the check (or attack).}

\item{pass}{Whether to calculate the DC necessary for passing the check (the
default) or failing it.}
}
\value{
A numeric scalar.
}
\description{
This is a simple function that calculates the required difficulty class so
that a skill check has the given success/failure probability. It hides the
complexities of the \link{roll} family of functions so users unfamiliar with R's
d/p/q/r notation can get quickly up and running with the package.
}
\details{
This function takes a roll expression (i.e., an arithmetic expression
involving dice) and calculates the DC of a skill check with a certain
success/failure probability. Since attacks and checks work in the same
way (i.e., roll a value higher than or equal to a certain threshold), there
is no \code{attack_dc()} function.

For more details on roll expressions, see \code{\link[=r]{r()}} and the \link{Dice} S4 class.
}
\examples{
# DC of a check so player has (at least) a 50\% chance of success
check_dc(d20 + 8, 0.5)

# DC of a check so player has (at most) a 30\% chance of failure
check_dc(d20 + 8, 0.3, pass = FALSE)

}
\seealso{
\link{roll}, \code{\link[=r]{r()}}, \link{Dice}.
}
