% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_prob}
\alias{check_prob}
\title{Get the probability of passing or failing a check}
\usage{
check_prob(roll, dc, pass = TRUE)
}
\arguments{
\item{roll}{A roll expression (e.g., \code{2 * d6 + 5}) or a data frame returned
by \code{\link[=r]{r()}}.}

\item{dc}{The dificulty class to beat for a skill check (or the armor class
to beat for an attack).}

\item{pass}{Whether to calculate the probability of passing the check (the
default) or failing it.}
}
\value{
A numeric scalar.
}
\description{
This is a simple function that calculates probabilities of hits/misses and
passes/fails. It hides the complexities of the \link{roll} family of functions so
users unfamiliar with R's d/p/q/r notation can get quickly up and running
with the package.
}
\details{
This function takes a roll expression (i.e., an arithmetic expression
involving dice) and calculates the probability of either hitting/missing an
attack or passing/failing a check. Since attacks and checks work in the same
way (i.e., roll a value higher than or equal to a certain threshold), there
is no \code{attack_hit()} function.

For more details on roll expressions, see \code{\link[=r]{r()}} and the \link{Dice} S4 class.
}
\examples{
# Probability of d20 + 8 passing a DC 15 skill check
check_prob(d20 + 8, 15)

# Probability of d20 + 8 missing an AC 15 attack
check_prob(d20 + 8, 15, pass = FALSE)

}
\seealso{
\link{roll}, \code{\link[=r]{r()}}, \link{Dice}.
}
