% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{d}
\alias{d}
\title{Create a Dice object}
\usage{
d(faces)
}
\arguments{
\item{faces}{A numeric vector with the die's faces.}
}
\value{
A Dice object.
}
\description{
Create a Dice object
}
\examples{
# Create a d6
d6 <- d(1:6)

}
