% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\docType{class}
\name{Dice-class}
\alias{Dice-class}
\title{An S4 class to represent dice}
\description{
An S4 class to represent dice
}
\section{Slots}{

\describe{
\item{\code{faces}}{A numeric vector with the die's faces.}
}}

