% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{Dice}
\alias{Dice}
\title{Create a Dice object}
\usage{
Dice(
  faces,
  weights = rep(1/length(faces), length(faces)),
  filter = function(x) x
)
}
\arguments{
\item{faces}{A numeric vector}

\item{weights}{A numeric vector}

\item{filter}{A function}
}
\description{
Create a Dice object
}
