% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{droll}
\alias{droll}
\title{Get relative frequency of an outcome of a roll}
\usage{
droll(x, roll)
}
\arguments{
\item{x}{A vector of outcomes.}

\item{roll}{A roll expression (e.g. \code{2 * d6 + 5}).}
}
\value{
A numeric vector.
}
\description{
Given a roll expression (i.e. an arithmetic expression involving dice),
compute the relative frequency of a specific outcome. In other words,
calculate the probability of obtaining this outcome.
}
\examples{
# Possible outcomes of 2d6 + 5
d6 <- d(1:6)
droll(7:9, 2 * d6 + 5)
}
