% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{droll_plot}
\alias{droll_plot}
\title{Plot the ways each outcome of a roll can be obtained}
\usage{
droll_plot(roll, ...)
}
\arguments{
\item{roll}{A roll expression (e.g. \code{2 * d6 + 5}).}

\item{...}{Other arguments passed on to \code{\link[graphics:barplot]{graphics::barplot()}}}
}
\value{
A bar plot.
}
\description{
Given a roll expression (i.e. an arithmetic expression involving dice),
compute the absolute frequency of each possible outcome and plot them as a
bar chart.
}
\examples{
# Possible outcomes of 2d6 + 5
d6 <- d(1:6)
droll_plot(2 * d6 + 5)
}
