% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{faces}
\alias{faces}
\alias{faces,Dice-method}
\title{Manipulate faces of a Dice object}
\usage{
faces(x)

\S4method{faces}{Dice}(x)
}
\arguments{
\item{x}{A Dice object.}
}
\description{
Manipulate faces of a Dice object
}
\section{Methods (by class)}{
\itemize{
\item \code{Dice}: Get faces of a Dice object.
}}

