% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{\%d\%}
\alias{\%d\%}
\title{Simple infix dice roller}
\usage{
n \%d\% sides
}
\arguments{
\item{n}{Number of dice.}

\item{sides}{Number of sides.}
}
\value{
A numeric scalar.
}
\description{
Simple infix dice roller
}
\examples{
# Roll 2d6
2 \%d\% 6

}
