% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{Math,Dice-method}
\alias{Math,Dice-method}
\title{Math operations with Dice objects}
\usage{
\S4method{Math}{Dice}(x)
}
\arguments{
\item{x}{A Dice object.}
}
\description{
Math operations with Dice objects
}
