% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{Math2,Dice-method}
\alias{Math2,Dice-method}
\title{Math2 operations with Dice objects}
\usage{
\S4method{Math2}{Dice}(x, digits)
}
\arguments{
\item{x}{A Dice object.}

\item{digits}{Number of digits to be used in \code{round} or \code{signif}.}
}
\description{
Math2 operations with Dice objects
}
