% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{Ops-Dice}
\alias{Ops-Dice}
\alias{Ops,Dice,numeric-method}
\alias{Ops,numeric,Dice-method}
\alias{Ops,Dice,Dice-method}
\alias{*,numeric,Dice-method}
\title{Arithmetic operations with Dice objects}
\usage{
\S4method{Ops}{Dice,numeric}(e1, e2)

\S4method{Ops}{numeric,Dice}(e1, e2)

\S4method{Ops}{Dice,Dice}(e1, e2)

\S4method{*}{numeric,Dice}(e1, e2)
}
\arguments{
\item{e1}{A numeric scalar or a Dice object.}

\item{e2}{A numeric scalar or a Dice object.}
}
\description{
Arithmetic operations with Dice objects
}
