% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{^,numeric,Dice-method}
\alias{^,numeric,Dice-method}
\title{Roll multiple dice and multiply outcomes}
\usage{
\S4method{^}{numeric,Dice}(e1, e2)
}
\arguments{
\item{e1}{A numeric value}

\item{e2}{A Dice object}
}
\description{
Roll multiple dice and multiply outcomes
}
