% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{proll}
\alias{proll}
\title{Get distribution of outcomes of a roll}
\usage{
proll(q, roll, lower.tail = TRUE)
}
\arguments{
\item{q}{A vector of outcomes.}

\item{roll}{A roll expression (e.g. \code{2 * d6 + 5}).}

\item{lower.tail}{Whether to return \verb{P[X ≤ x]} or \code{P[X > x]}.}
}
\value{
A numeric vector.
}
\description{
Given a roll expression (i.e. an arithmetic expression involving dice),
compute the distribution the outcomes, returning \code{P[X <= x]} where \code{x} is an
outcome of the roll.
}
\examples{
# Possible outcomes of 2d6 + 5
d6 <- d(1:6)
proll(7:9, 2 * d6 + 5)
}
