% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{proll_plot}
\alias{proll_plot}
\title{Plot probability distribution}
\usage{
proll_plot(roll, ..., lower.tail = TRUE)
}
\arguments{
\item{roll}{A roll expression (e.g. \code{2 * d6 + 5}).}

\item{...}{Other arguments passed on to \code{\link[graphics:barplot]{graphics::barplot()}}}

\item{lower.tail}{Whether to return \code{P[X <= x]} or \code{P[X > x]}.}
}
\value{
A bar plot.
}
\description{
Plot probability distribution
}
\examples{
# Possible outcomes of 2d6 + 5
d6 <- d(1:6)
proll_plot(2 * d6 + 5)
}
