% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{qroll_plot}
\alias{qroll_plot}
\title{Plot quantile distribution}
\usage{
qroll_plot(roll, ..., lower.tail = TRUE)
}
\arguments{
\item{roll}{A roll expression (e.g. \code{2 * d6 + 5}).}

\item{...}{Other arguments passed on to \code{\link[graphics:barplot]{graphics::barplot()}}}

\item{lower.tail}{Whether to return \code{x | P[X <= x] >= p} or
\code{x | P[X > x] >= p}.}
}
\value{
A bar plot.
}
\description{
Plot quantile distribution
}
\examples{
# Possible outcomes of 2d6 + 5
d6 <- d(1:6)
qroll_plot(2 * d6 + 5)
}
