% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{r}
\alias{r}
\alias{r,Dice-method}
\title{Simulate rolling a die}
\usage{
r(object, n = 1)

\S4method{r}{Dice}(object, n = 1)
}
\arguments{
\item{object}{A Dice object.}

\item{n}{Number of dice.}
}
\description{
Simulate rolling a die
}
\section{Methods (by class)}{
\itemize{
\item \code{Dice}: Roll a Dice object.
}}

