% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{roll_outcome_count_}
\alias{roll_outcome_count_}
\title{Count the ways each outcome of a roll can be obtained (internal)}
\usage{
roll_outcome_count_(roll, env = parent.frame())
}
\arguments{
\item{roll}{A roll expression (e.g. \code{2 * d6 + 5}).}

\item{env}{The environment of \code{roll}.}
}
\value{
A data frame with two columns: \code{outcome} and \code{count}.
}
\description{
Given a roll expression, compute the absolute frequency of each possible
outcome. In other words, calculate how many ways can every outcome of the
roll be obtained. \code{\link[=roll_outcome_count]{roll_outcome_count()}} wraps this function so that the user
doesn't have to worry about environments.
}
