% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{roll_outcome_freq}
\alias{roll_outcome_freq}
\title{Calculate the frequency with which each outcome of a roll can be obtained}
\usage{
roll_outcome_freq(roll, precise = FALSE)
}
\arguments{
\item{roll}{A roll expression (e.g. \code{2 * d6 + 5}).}

\item{precise}{Whether to return values with arbitrary precision.}
}
\value{
A data frame with two columns: \code{outcome} and \code{freq}.
}
\description{
Given a roll expression (i.e. an arithmetic expression involving dice),
compute the relative frequency of each possible outcome. In other words,
calculate how many ways every outcome of the roll can be obtained and then
divide by the total sum.
}
\examples{
# Possible outcomes of 2d6 + 5
d6 <- d(1:6)
roll_outcome_freq(2 * d6 + 5)
}
