% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rroll_plot}
\alias{rroll_plot}
\title{Plot random samples of a roll}
\usage{
rroll_plot(n, roll, ...)
}
\arguments{
\item{n}{Number of samples.}

\item{roll}{A roll expression (e.g. \code{2 * d6 + 5}).}

\item{...}{Other arguments passed on to \code{\link[graphics:barplot]{graphics::barplot()}}}
}
\value{
A histogram.
}
\description{
Plot random samples of a roll
}
\examples{
# Possible outcomes of 2d6 + 5
d6 <- d(1:6)
rroll_plot(1000, 2 * d6 + 5)
}
