% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{Summary,Dice-method}
\alias{Summary,Dice-method}
\title{Summary operations with Dice objects}
\usage{
\S4method{Summary}{Dice}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A Dice object}

\item{...}{Numeric arguments.}

\item{na.rm}{A logical indicating whether missing values should be removed.}
}
\description{
Summary operations with Dice objects
}
