% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{*,numeric,Dice-method}
\alias{*,numeric,Dice-method}
\title{Simulate rolling multiple Dice objects and add outcomes}
\usage{
\S4method{*}{numeric,Dice}(e1, e2)
}
\arguments{
\item{e1}{A numeric scalar.}

\item{e2}{A Dice object.}
}
\description{
Simulate rolling multiple Dice objects and add outcomes
}
