% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{droll}
\alias{droll}
\title{Count the ways each outcome of a roll can be obtained}
\usage{
droll(roll, precise = FALSE)
}
\arguments{
\item{roll}{A roll expression (e.g. \code{2 * d6 + 5}).}

\item{precise}{Whether to return values with arbitrary precision.}
}
\value{
A data frame with two columns: \code{outcome} and \code{count}.
}
\description{
Given a roll expression (i.e. an arithmetic expression involving dice),
compute the absolute frequency of each possible outcome. In other words,
calculate how many ways every outcome of the roll can be obtained.
}
\examples{
# Possible outcomes of 2d6 + 5
d6 <- d(1:6)
droll(2 * d6 + 5)
}
