% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{qroll}
\alias{qroll}
\title{Get inverse distribution of outcomes of a roll}
\usage{
qroll(p, roll, lower.tail = TRUE)
}
\arguments{
\item{p}{A vector of probabilities.}

\item{roll}{A roll expression (e.g. \code{2 * d6 + 5}).}

\item{lower.tail}{Whether to return \verb{x | P[X ≤ x] = p} or \code{x | P[X > x] = p}.}
}
\value{
A numeric vector.
}
\description{
Given a roll expression (i.e. an arithmetic expression involving dice),
compute the distribution the outcomes, returning \verb{x | P[X ≤ x] = p} where
\code{p} is a probability.
}
\examples{
# Possible outcomes of 2d6 + 5
d6 <- d(1:6)
qroll(7:9, 2 * d6 + 5)
}
