% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll some dice}
\usage{
roll(roll, verbose = TRUE)
}
\arguments{
\item{roll}{A roll expression (e.g. \code{2 * d6 + 5}).}

\item{verbose}{Whether to print the expression after the dice rolls.}
}
\value{
A numeric scalar.
}
\description{
Given a roll expression (i.e. an arithmetic expression involving dice),
compute an outcome by sampling one face from each die and evaluating the
expression. If \code{verbose = TRUE}, also print the expression before evaluation.
}
\examples{
# Roll 2d6 + 6
d6 <- d(1:6)
roll(2 * d6 + 5)

}
