% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{roll_outcome_count}
\alias{roll_outcome_count}
\title{Calculate the absolute frequency of each outcome of a roll formula}
\usage{
roll_outcome_count(roll)
}
\arguments{
\item{roll}{Roll formula}
}
\description{
Calculate the absolute frequency of each outcome of a roll formula
}
